requirements.txt
numpy==1.20.3
matplotlib==3.5.0
pillow==8.3.1
pytorch==1.9.0
torchvision==0.10.0
tensorboard==2.9.1
opencv-python==4.6.0.66
opencv_contrib_python==4.6.0.66
urllib3==1.25.8
scipy==1.8.1

